requirements.txt
django==1.5
django-registration==0.8
python-ldap==2.4.10
PIL==1.1.7
